refactor: Custom Hook 생성
